% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{add_dimension}
\alias{add_dimension}
\title{Add a Dummy Dimension to an array}
\usage{
add_dimension(array)
}
\arguments{
\item{array}{any array}
}
\value{
a array with an added dummy dimension
}
\description{
Adds a nominal dimension to the array to conform with Keras standards
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
