% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{build_model}
\alias{build_model}
\title{Build Keras Convolutional Neural Network}
\usage{
build_model()
}
\value{
an untrained Keras CNN with randomized params
}
\description{
Constructs a Keras NN with 4 convolutional layers
}
\note{
it might be necessary to run install_packages first
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
