% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{check_imageset}
\alias{check_imageset}
\title{Dimension check an image set}
\usage{
check_imageset(imageset)
}
\arguments{
\item{imageset}{a set of images bound along axis 1}
}
\value{
a set of images bound along axis 1 with appropriate dimensions (128 by 64)
}
\description{
Check for the correct dimensions
}
\examples{
df = array(c(100, 500, 300))
new_df = check_imageset(df)

}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
