% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{check_keys}
\alias{check_keys}
\title{Dimension check a key array}
\usage{
check_keys(keys)
}
\arguments{
\item{keys}{an array containing labels for each image in an imageset, bound along 1st dimension}
}
\value{
keys array in the proper form
}
\description{
Binds input with a negated self if input is 1-dimensional logical vector
}
\examples{
one_dim_key = array(0,c(10,1))
two_dim_key = check_keys(one_dim_key)

}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
