% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{install_packages}
\alias{install_packages}
\title{Install Keras, Tensorflow and Supporting Packages}
\usage{
install_packages()
}
\value{
None
}
\description{
Installs packages necessary to make and run neural networks
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
