% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{load_model}
\alias{load_model}
\title{Load Model from File}
\usage{
load_model(filename)
}
\arguments{
\item{filename}{a string matching the name of the target file}
}
\value{
keras model
}
\description{
Loads a model saved with the Keras standard
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
