% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{load_premade_data}
\alias{load_premade_data}
\title{Load Sample ImageSet and Keys}
\usage{
load_premade_data(n_samples)
}
\arguments{
\item{n_samples}{max of 400; number of samples to be produced in the data set}
}
\value{
an N by 128 by 64 imagest and N by 2 key bound into a list
}
\description{
Loads a premade set of data based on a nutrient-phytoplankton model
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
