% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{load_premade_model}
\alias{load_premade_model}
\title{Load Pre-Trained Model from File}
\usage{
load_premade_model(type)
}
\arguments{
\item{type}{either classify by "domain" or identify if there is a "transition"}
}
\value{
a pre-trained keras model
}
\description{
Loads a specific model from a bank of pre-trained models
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
