% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{package_check}
\alias{package_check}
\title{Check if a package is present}
\usage{
package_check(x)
}
\value{
None
}
\description{
Install package if not present
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
