% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{predict}
\alias{predict}
\title{Predict 1-Channel Image Classification}
\usage{
predict(model, imageset, type = "category")
}
\arguments{
\item{model}{a pre-trained CNN using the Keras package}

\item{imageset}{an N by 128 by 64 dimensional array; N is the number of different imageset}

\item{type}{either "category" for binary classsification or "probability" for category likelihoods}
}
\value{
N length vector containing a binary prediction of classification
}
\description{
Uses a pre-trained model to classify an image or set of images
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
