% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{prob_event2d}
\alias{prob_event2d}
\title{Predict Probability of a 2-Variable Event}
\usage{
prob_event2d(event, other_events)
}
\arguments{
\item{event}{a vector of length two}

\item{other_events}{a two column matrix of paired values}
}
\value{
probability of that event occuring with the other events
}
\description{
Predicts the probability ofa 2D outcome given all other outcomes
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
