% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{prob_timeseries}
\alias{prob_timeseries}
\title{Predict Probability of Time Series}
\usage{
prob_timeseries(timeseries2var)
}
\arguments{
\item{timeseries2var}{a two column matrix of two-variable events}
}
\value{
vector of probabilities for each of the time series values
}
\description{
Calculates probability of each point in the time series given all previous points
}
\note{
for forecasting, you should be using predict() outputs, not classify() outputs
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
