% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{resize_images}
\alias{resize_images}
\title{Resize Images to Model Standard}
\usage{
resize_images(imageset)
}
\arguments{
\item{imageset}{an N by x by y dimensional array; N is the number of different images}
}
\value{
N by 128 by 64 dimensional array
}
\description{
Takes a set of images concatenated along the 1st dimension and resizes them to 128 by 64
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
