% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{save_model}
\alias{save_model}
\title{Save Model to File}
\usage{
save_model(model, filename)
}
\arguments{
\item{model}{a keras model to be saved}

\item{filename}{a string matching the name of the target file}
}
\description{
Save a model saved with the Keras standard
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
