% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{train_model}
\alias{train_model}
\title{March 18
Produce trained model from images and their keys}
\usage{
train_model(
  imageset,
  keys,
  model = NULL,
  test_iterations = 500,
  train_test_ratio = 0.7,
  batch_size = 128
)
}
\arguments{
\item{imageset}{an N x W x H set of images with features for classification training}

\item{keys}{an N by 2 logical matrix/array of 2-value keys corresponding to features in imageset}

\item{model}{optionally input a pre-built model for more training}
}
\value{
a trained cnn
}
\description{
Train a convolutional neural network from a labeled data set
}
\note{
not tested with non-logical keys
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
