% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{wt}
\alias{wt}
\title{Wavelet Transform a Timeseries}
\usage{
wt(timeseries, key = NULL)
}
\arguments{
\item{timeseries}{a one-variable timeseries in vector form}

\item{key}{optional input is wrapped in list with output wavelet transform}
}
\value{
128 by 64 matrix of wavelet power OR a list containing 1: wavelet matrix and 2: the input key
}
\description{
Gives the 2D wavelet transform of a one variable timeseries
}
\note{
inputting a key changes the output from matrix to list
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
