% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-UserFunctions.R
\name{wt_windowed}
\alias{wt_windowed}
\title{Wavelet Transform a Timeseries}
\usage{
wt_windowed(timeseries, window_width, window_stride, shift_index = NULL)
}
\arguments{
\item{timeseries}{a one-variable timeseries in vector form}

\item{window_width}{width of analysis window measured in number of timeseries indices}

\item{window_stride}{stride of analysis window measured in number of timeseries indices}

\item{shift_index}{optional input, will create a vector indicator for windows before and after the index with 0s and 1s, respectively}
}
\value{
128 by 64 matrix of wavelet power OR a list with 1: wavelet matrix and 2: key vector
}
\description{
Gives the 2D wavelet transform of a one variable timeseries
}
\author{
Ryan Taylor \email{rmt2dt@virginia.edu}, Jonathan Walter
\email{jaw3es@virginia.edu}
}
